pyproject.toml
# ABOUTME: Package configuration for ethica CLI tool
# ABOUTME: Defines dependencies, build system, and development tools

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ethica"
version = "0.1.0"
description = "Framework-agnostic platform for AI ethics compliance checking"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Ethics AI Commons Contributors"}
]
keywords = ["ai", "ethics", "compliance", "unesco", "responsible-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-pyyaml",
]

[project.scripts]
ethica = "ethica.__main__:app"

[project.urls]
Homepage = "https://github.com/shellen/ethica"
Documentation = "https://github.com/shellen/ethica/blob/main/docs/getting-started.md"
Repository = "https://github.com/shellen/ethica"
Issues = "https://github.com/shellen/ethica/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ethica*"]

[tool.setuptools.package-data]
ethica = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ethica --cov-report=term-missing"
